requirements.txt
PyMuPDF 
pytesseract 
Pillow 
chainlit
langchain
chromadb
PyPDF2
tiktoken
io
fitz